test(cli): split star/unstar/stars tests into separate files
